APPLE: various fixes
