+++ nym-vpn-apple/Home/Sources/26/OneClick/OneClickViewModel.swift
@@ -20,15 +20,16 @@ import GRPCManager
 @MainActor
 @Observable
 public final class OneClickViewModel {
-    private let appSettings: AppSettings
-    private let connectionManager: ConnectionManager
-    private let credentialsManager: CredentialsManager
-    private let gatewayManager: GatewayManager
-    private let snackbarManager: SnackbarManager
-    private let impactGenerator: ImpactGenerator
-    private let networkMonitor: NetworkMonitor
+    // Internal (not `private`) so the `+State` / `+Selection` extension files can reach them.
+    let appSettings: AppSettings
+    let connectionManager: ConnectionManager
+    let credentialsManager: CredentialsManager
+    let gatewayManager: GatewayManager
+    let snackbarManager: SnackbarManager
+    let impactGenerator: ImpactGenerator
+    let networkMonitor: NetworkMonitor
 #if os(macOS)
-    private let grpcManager: GRPCManager
+    let grpcManager: GRPCManager
 #endif
 
     var connectState: OneClickConnectState = .disconnected
@@ -55,10 +56,10 @@ public final class OneClickViewModel {
     /// daemon is not running, so the user can install/enable it.
     @ObservationIgnored public var onRequestDaemonEnable: (() -> Void)?
 
-    @ObservationIgnored private var connectDisconnectTask: Task<Void, Never>?
-    @ObservationIgnored private var resolveTask: Task<Void, Never>?
-    @ObservationIgnored private var cancellables = Set<AnyCancellable>()
-    @ObservationIgnored private var isConnectDisconnectInFlight = false
+    @ObservationIgnored var connectDisconnectTask: Task<Void, Never>?
+    @ObservationIgnored var resolveTask: Task<Void, Never>?
+    @ObservationIgnored var cancellables = Set<AnyCancellable>()
+    @ObservationIgnored var isConnectDisconnectInFlight = false
 
 #if os(iOS)
     public init(
@@ -127,53 +128,7 @@ public final class OneClickViewModel {
 
         connectDisconnectTask?.cancel()
         connectDisconnectTask = Task { @MainActor [weak self] in
-            guard let self else { return }
-            isConnectDisconnectInFlight = true
-            defer { isConnectDisconnectInFlight = false }
-
-            if isConnectingTap {
-#if os(iOS)
-                if !networkMonitor.isAvailable {
-                    presentOfflineAlert()
-                    return
-                }
-#endif
-#if os(macOS)
-                if !grpcManager.isServing {
-                    onRequestDaemonEnable?()
-                    return
-                }
-#endif
-                guard credentialsManager.isValidCredentialImported else { return }
-                if await !credentialsManager.isAccountValid() {
-                    await credentialsManager.updateAccountSummary()
-                    let summary = credentialsManager.accountSummary
-                    let shouldOfferPurchase = ConnectPlanPurchaseGatePolicy.shouldOfferPlanPurchaseOnConnect(
-                        isAccountRegistrationInFlight: credentialsManager.isAccountRegistrationInFlight,
-                        accountSummaryLastFetchFailed: credentialsManager.accountSummaryLastFetchFailed,
-                        isAccountActive: credentialsManager.isAccountActive(),
-                        validUntilIsFuture: LoginSessionPolicy.validUntilIsFuture(
-                            validUntil: summary?.validUntilDate
-                        ),
-                        hasAccountSummary: summary != nil
-                    )
-                    if shouldOfferPurchase {
-                        sessionCoordinator?.handle(.requestInactiveSubscriptionPurchase)
-                        return
-                    }
-                }
-            }
-
-            do {
-                try await connectionManager.connectDisconnect()
-            } catch {
-                impactGenerator.error()
-                presentConnectionErrorAlert(
-                    message: ConnectionStatusViewModel.userFacingMessage(from: error)
-                )
-            }
-            handleInactiveSubscriptionErrorIfNeeded()
-            clearLastErrorIfNeeded()
+            await self?.performConnectDisconnect(isConnectingTap: isConnectingTap)
         }
     }
 
@@ -182,14 +137,10 @@ public final class OneClickViewModel {
         connectDisconnectTask = Task { @MainActor [weak self] in
             guard let self else { return }
             snackbarManager.clear()
-            guard connectionManager.shouldDisconnectActiveTunnel() else {
-                connectionManager.lastError = nil
-                return
-            }
             isConnectDisconnectInFlight = true
             defer { isConnectDisconnectInFlight = false }
             do {
-                try await connectionManager.disconnectActiveTunnel()
+                try await connectionManager.connectDisconnect()
                 connectionManager.lastError = nil
             } catch {
                 impactGenerator.error()
@@ -221,22 +172,6 @@ public final class OneClickViewModel {
 
         impactGenerator.impact()
         snackbarManager.clear()
-
-        connectDisconnectTask?.cancel()
-        connectDisconnectTask = Task { @MainActor [weak self] in
-            guard let self else { return }
-            isConnectDisconnectInFlight = true
-            defer { isConnectDisconnectInFlight = false }
-            do {
-                try await connectionManager.connectDisconnect()
-            } catch {
-                impactGenerator.error()
-                presentConnectionErrorAlert(
-                    message: ConnectionStatusViewModel.userFacingMessage(from: error)
-                )
-            }
-            clearLastErrorIfNeeded()
-        }
     }
 
     func upCaretTapped() {
@@ -275,463 +210,7 @@ public final class OneClickViewModel {
     }
 }
 
-private extension OneClickViewModel {
-    func seedFromCurrentValues() {
-        recomputeConnectState()
-        refreshSelection()
-    }
-
-    func observe() {
-        connectionManager.$currentTunnelStatus
-            .removeDuplicates()
-            .receive(on: DispatchQueue.main)
-            .sink { [weak self] status in
-                guard let self else { return }
-                if case .connected = status {
-                    impactGenerator.success()
-                }
-                recomputeConnectState()
-                refreshSelection()
-            }
-            .store(in: &cancellables)
-
-        connectionManager.$lastError
-            .receive(on: DispatchQueue.main)
-            .sink { [weak self] _ in
-                self?.recomputeConnectState()
-            }
-            .store(in: &cancellables)
-
-        connectionManager.$connectionInfoData
-            .receive(on: DispatchQueue.main)
-            .sink { [weak self] _ in
-                self?.refreshSelection()
-            }
-            .store(in: &cancellables)
-
-        connectionManager.$connectionConfig
-            .receive(on: DispatchQueue.main)
-            .sink { [weak self] config in
-                guard let self else { return }
-                speedMode = OneClickSpeedMode(isTwoHop: config.enableTwoHop)
-                refreshSelection()
-            }
-            .store(in: &cancellables)
-
-#if SANTA
-        Publishers.MergeMany(
-            gatewayManager.$entry.map { _ in () },
-            gatewayManager.$exit.map { _ in () },
-            gatewayManager.$vpn.map { _ in () }
-        )
-        .receive(on: DispatchQueue.main)
-        .sink { [weak self] in
-            self?.refreshSelection()
-        }
-        .store(in: &cancellables)
-#endif
-
-        credentialsManager.$accountSummary
-            .receive(on: DispatchQueue.main)
-            .sink { [weak self] _ in
-                self?.recomputeConnectState()
-            }
-            .store(in: &cancellables)
-
-#if os(iOS)
-        networkMonitor.$isAvailable
-            .receive(on: DispatchQueue.main)
-            .sink { [weak self] _ in
-                self?.recomputeConnectState()
-            }
-            .store(in: &cancellables)
-#endif
-
-#if os(macOS)
-        grpcManager.$isServing
-            .receive(on: DispatchQueue.main)
-            .sink { [weak self] _ in
-                self?.recomputeConnectState()
-            }
-            .store(in: &cancellables)
-#endif
-    }
-
-    func recomputeConnectState() {
-        let next = derivedConnectState()
-        guard next != connectState else { return }
-        connectState = next
-    }
-
-    func derivedConnectState() -> OneClickConnectState {
-        switch connectionManager.currentTunnelStatus {
-        case .connected:
-            return .connected
-        case .disconnecting:
-            return .disconnecting
-        case .error:
-            return .stop
-        case .connecting, .reasserting, .restarting, .offlineReconnect:
-            return .connecting
-        case .disconnected, .offline, .unknown:
-#if os(iOS)
-            if !networkMonitor.isAvailable {
-                return .noInternet
-            }
-#endif
-            if credentialsManager.isValidCredentialImported, !credentialsManager.isAccountActive() {
-                return .noSubscription
-            }
-            return .disconnected
-        }
-    }
-
-    func handleInactiveSubscriptionErrorIfNeeded() {
-        guard connectionManager.currentTunnelStatus == .error else { return }
-        guard let error = connectionManager.lastError else { return }
-        let reason: ErrorReason?
-        if let typed = error as? ErrorReason {
-            reason = typed
-        } else {
-            let nsError = error as NSError
-            reason = nsError.domain == ErrorReason.domain ? ErrorReason(nsError: nsError) : nil
-        }
-        guard reason == .inactiveSubscription else { return }
-        sessionCoordinator?.handle(.requestInactiveSubscriptionPurchase)
-    }
-
-    func clearLastErrorIfNeeded() {
-        if isAwaitingGatewayIndependenceConsent {
-            return
-        }
-        switch connectionManager.currentTunnelStatus {
-        case .disconnecting, .disconnected, .error:
-            connectionManager.lastError = nil
-        default:
-            break
-        }
-    }
-
-    var isAwaitingGatewayIndependenceConsent: Bool {
-        GatewayIndependenceArcPolicy.shouldPreserveIndependenceConsentError(
-            status: connectionManager.currentTunnelStatus,
-            lastError: connectionManager.lastError
-        )
-    }
-
-    func applyDisplayMode(_ mode: OneClickDisplayMode) {
-        displayMode = mode
-        appSettings.oneClickDisplayModeRaw = mode.rawValue
-        refreshSelection()
-    }
-
-    func refreshSelection() {
-        let cfg = connectionManager.connectionConfig
-        let twoHop = cfg.enableTwoHop
-        let exitType: NodeType = twoHop ? .vpn : .exit
-        let entryType: NodeType = twoHop ? .vpn : .entry
-        let info = connectionManager.connectionInfoData
-        let liveCandidate = isLiveStatus(connectionManager.currentTunnelStatus) && info != nil
-        resolveTask?.cancel()
-        resolveTask = Task { @MainActor [weak self] in
-            guard let self else { return }
-            if liveCandidate,
-               let info,
-               let exitLive = livePhase(
-                   gatewayId: info.exitGatewayId,
-                   gatewayType: exitType,
-                   hopType: .exit,
-                   displayMode: displayMode
-               ) {
-                isLiveConnection = true
-                selectionPhase = exitLive
-                if displayMode == .nerd {
-                    entrySelectionPhase = livePhase(
-                        gatewayId: info.entryGatewayId,
-                        gatewayType: entryType,
-                        hopType: .entry,
-                        displayMode: displayMode
-                    )
-                        ?? resolveEntryPhase(
-                            entry: cfg.entry,
-                            gatewayId: info.entryGatewayId ?? cfg.entry.gatewayId,
-                            gatewayType: entryType
-                        )
-                } else {
-                    entrySelectionPhase = .selecting
-                }
-                return
-            }
-            isLiveConnection = false
-            switch displayMode {
-            case .powerUser:
-                entrySelectionPhase = .selecting
-                selectionPhase = resolveExitPhase(
-                    exit: cfg.exit,
-                    gatewayId: cfg.exit.gatewayId,
-                    gatewayType: exitType
-                )
-            case .nerd:
-                entrySelectionPhase = resolveEntryPhase(
-                    entry: cfg.entry,
-                    gatewayId: cfg.entry.gatewayId,
-                    gatewayType: entryType
-                )
-                selectionPhase = resolveExitPhase(
-                    exit: cfg.exit,
-                    gatewayId: cfg.exit.gatewayId,
-                    gatewayType: exitType
-                )
-            }
-        }
-    }
-
-    func isLiveStatus(_ status: TunnelStatus) -> Bool {
-        switch status {
-        case .connecting, .reasserting, .restarting, .connected, .offlineReconnect, .disconnecting:
-            return true
-        case .disconnected, .offline, .error, .unknown:
-            return false
-        }
-    }
-
-    func livePhase(
-        gatewayId: String?,
-        gatewayType: NodeType,
-        hopType: HopType,
-        displayMode: OneClickDisplayMode
-    ) -> OneClickSelectionPhase? {
-        guard let gatewayId,
-              let gateway = gatewayManager.gateway(with: gatewayId, gatewayType: gatewayType)
-        else { return nil }
-        let title: String
-        switch displayMode {
-        case .nerd:
-            title = gateway.name ?? gateway.id
-        case .powerUser:
-            title = gateway.ipv4s.first ?? gateway.name ?? gateway.id
-        }
-        let location = gateway.location
-        let subtitle: String?
-        if let location {
-            let country = gatewayManager.localizedCountry(with: location.twoLetterIsoCountryCode)?.name
-            if gatewayManager.shouldDisplayRegion(with: location.twoLetterIsoCountryCode) {
-                subtitle = "\(location.city), \(location.region), \(country ?? "")"
-            } else {
-                subtitle = "\(location.city), \(country ?? "")"
-            }
-        } else {
-            subtitle = nil
-        }
-        return .selected(OneClickServerInfo(
-            countryCode: location?.twoLetterIsoCountryCode ?? "",
-            title: title,
-            subtitle: subtitle,
-            score: score(for: gateway),
-            gateway: gateway,
-            hopType: hopType,
-            showsInfoButton: true
-        ))
-    }
-
-    func resolveEntryPhase(entry: EntryGateway, gatewayId: String?, gatewayType: NodeType) -> OneClickSelectionPhase {
-        let gateway = gatewayId.flatMap { gatewayManager.gateway(with: $0, gatewayType: gatewayType) }
-        let title: String
-        if case .random = entry {
-            title = "gatewaysView.random".localizedString
-        } else {
-            title = gatewayManager.userFriendlyTitle(with: entry)
-                ?? gateway?.location?.city ?? gateway?.name ?? gateway?.id ?? ""
-        }
-        let subtitle = entrySubtitle(entry: entry, gateway: gateway)
-        let countryCode = gatewayManager.countryCode(with: entry)
-            ?? gateway?.location?.twoLetterIsoCountryCode ?? ""
-        let scoreGateway = scoreGateway(forEntry: entry, fallback: gateway, gatewayType: gatewayType)
-        let isExplicitGateway: Bool
-        if case .gateway = entry { isExplicitGateway = true } else { isExplicitGateway = false }
-        return makePhase(
-            title: title,
-            subtitle: subtitle,
-            countryCode: countryCode,
-            scoreGateway: scoreGateway,
-            hopType: .entry,
-            showsInfoButton: isExplicitGateway
-        )
-    }
-
-    func resolveExitPhase(exit: ExitRouter, gatewayId: String?, gatewayType: NodeType) -> OneClickSelectionPhase {
-        let gateway = gatewayId.flatMap { gatewayManager.gateway(with: $0, gatewayType: gatewayType) }
-        let title: String
-        if case .random = exit {
-            title = "gatewaysView.random".localizedString
-        } else {
-            title = gatewayManager.userFriendlyTitle(with: exit)
-                ?? gateway?.location?.city ?? gateway?.name ?? gateway?.id ?? ""
-        }
-        let subtitle = exitSubtitle(exit: exit, gateway: gateway)
-        let countryCode = gatewayManager.countryCode(with: exit)
-            ?? gateway?.location?.twoLetterIsoCountryCode ?? ""
-        let scoreGateway = scoreGateway(forExit: exit, fallback: gateway, gatewayType: gatewayType)
-        let isExplicitGateway: Bool
-        if case .gateway = exit { isExplicitGateway = true } else { isExplicitGateway = false }
-        return makePhase(
-            title: title,
-            subtitle: subtitle,
-            countryCode: countryCode,
-            scoreGateway: scoreGateway,
-            hopType: .exit,
-            showsInfoButton: isExplicitGateway
-        )
-    }
-
-    func scoreGateway(forEntry entry: EntryGateway, fallback: GatewayNode?, gatewayType: NodeType) -> GatewayNode? {
-        switch entry {
-        case .country, .region, .random:
-            return gatewayManager.bestGateway(matching: entry, gatewayType: gatewayType) ?? fallback
-        case .gateway:
-            return fallback
-        }
-    }
-
-    func scoreGateway(forExit exit: ExitRouter, fallback: GatewayNode?, gatewayType: NodeType) -> GatewayNode? {
-        switch exit {
-        case .country, .region, .random:
-            return gatewayManager.bestGateway(matching: exit, gatewayType: gatewayType) ?? fallback
-        case .gateway:
-            return fallback
-        }
-    }
-
-    func makePhase(
-        title: String,
-        subtitle: String?,
-        countryCode: String,
-        scoreGateway: GatewayNode?,
-        hopType: HopType,
-        showsInfoButton: Bool = false
-    ) -> OneClickSelectionPhase {
-        if title.isEmpty && (subtitle?.isEmpty ?? true) {
-            return .selecting
-        }
-        return .selected(OneClickServerInfo(
-            countryCode: countryCode,
-            title: title,
-            subtitle: subtitle,
-            score: scoreGateway.map { score(for: $0) } ?? .offline,
-            gateway: scoreGateway,
-            hopType: hopType,
-            showsInfoButton: showsInfoButton
-        ))
-    }
-
-    func score(for gateway: GatewayNode) -> OneClickServerScore {
-        let raw: GatewayNodeScore?
-        switch connectionManager.connectionType {
-        case .wireguard:
-            raw = gateway.performance?.score
-        case .mixnet5hop:
-            raw = gateway.performance?.mixnetScore ?? gateway.mixnetScore
-        }
-        return Self.score(from: raw ?? .offline)
-    }
-
-    func entrySubtitle(entry: EntryGateway, gateway: GatewayNode?) -> String? {
-        guard let location = gateway?.location else { return nameOrId(gateway: gateway) }
-        switch entry {
-        case let .country(code):
-            return countrySubtitle(gateway: gateway, countryCode: code, location: location)
-        case .region:
-            return regionSubtitle(gateway: gateway, location: location)
-        case .gateway:
-            return serverSubtitle(location: location, countryCode: location.twoLetterIsoCountryCode)
-        case .random:
-            return nil
-        }
-    }
-
-    func exitSubtitle(exit: ExitRouter, gateway: GatewayNode?) -> String? {
-        guard let location = gateway?.location else { return nameOrId(gateway: gateway) }
-        switch exit {
-        case let .country(code):
-            return countrySubtitle(gateway: gateway, countryCode: code, location: location)
-        case .region:
-            return regionSubtitle(gateway: gateway, location: location)
-        case .gateway:
-            return serverSubtitle(location: location, countryCode: location.twoLetterIsoCountryCode)
-        case .random:
-            return nil
-        }
-    }
-
-    func countrySubtitle(gateway: GatewayNode?, countryCode: String, location: GatewayNodeLocation) -> String? {
-        if gatewayManager.shouldDisplayRegion(with: countryCode) {
-            "\(location.city), \(location.region) \(nameOrId(gateway: gateway))"
-        } else {
-            "\(location.city) \(nameOrId(gateway: gateway))"
-        }
-    }
-
-    func regionSubtitle(gateway: GatewayNode?, location: GatewayNodeLocation) -> String? {
-        "\(location.city) \(nameOrId(gateway: gateway))"
-    }
-
-    func serverSubtitle(location: GatewayNodeLocation, countryCode: String) -> String? {
-        let country = gatewayManager.localizedCountry(with: countryCode)
-        if gatewayManager.shouldDisplayRegion(with: countryCode) {
-            return "\(location.city), \(location.region), \(country?.name ?? "")"
-        } else {
-            return "\(location.city), \(country?.name ?? "")"
-        }
-    }
-
-    func nameOrId(gateway: GatewayNode?) -> String {
-        if let name = gateway?.name {
-            "(\(name))"
-        } else if let identifier = gateway?.id {
-            "(\(identifier))"
-        } else {
-            ""
-        }
-    }
-
-    static func score(from node: GatewayNodeScore) -> OneClickServerScore {
-        switch node {
-        case .high:
-            return .high
-        case .medium:
-            return .medium
-        case .low:
-            return .low
-        case .offline, .noScore:
-            return .offline
-        }
-    }
-
-    func presentOfflineAlert() {
-        snackbarManager.enqueue(
-            SnackbarItem(
-                style: .warning,
-                title: "home.modal.noInternetConnection.title".localizedString,
-                message: "home.modal.noInternetConnection.subtitle".localizedString
-            )
-        )
-    }
-
-    func presentConnectionErrorAlert(message: String) {
-        snackbarManager.enqueue(
-            SnackbarItem(
-                style: .critical,
-                title: "connectionError.title".localizedString,
-                message: ConnectionErrorCopy.message(reason: message),
-                actionTitle: "disconnect".localizedString,
-                onAction: { [weak self] in self?.disconnectFromError() },
-                duration: 7
-            )
-        )
-    }
-}
-
-private extension OneClickSpeedMode {
+extension OneClickSpeedMode {
     init(isTwoHop: Bool) {
         self = isTwoHop ? .fast : .anonymous
     }
