+++ nym-vpn-apple/NymMixnetTunnel/PacketTunnelProvider.swift
@@ -48,7 +48,9 @@ class PacketTunnelProvider: NEPacketTunnelProvider {
         let vpnConfig = try mixnetConfig.asVpnConfig(tunProvider: self)
         try await setup(vpnConfig: vpnConfig)
 
-        try await ensureGatewayIndependenceAllowsConnect()
+        try await ensureGatewayIndependenceAllowsConnect(
+            remindersEnabled: mixnetConfig.isServerFamilyRemindersEnabled
+        )
 
         _ = try await commandSender?.connectTunnel()
         try await tunnelActor.waitUntilStarted()
@@ -70,8 +72,16 @@ class PacketTunnelProvider: NEPacketTunnelProvider {
 }
 
 extension PacketTunnelProvider {
-    func ensureGatewayIndependenceAllowsConnect() async throws {
+    func ensureGatewayIndependenceAllowsConnect(remindersEnabled: Bool) async throws {
         guard let commandSender else { return }
+
+        // Reminders off means the user opted out of the warning: relax up front
+        // so no `needsRelaxedIndependenceCriteria` error (and no modal) can surface.
+        guard remindersEnabled
+        else {
+            try await commandSender.setEnableGatewayIndependence(enableGatewayIndependence: false)
+            return
+        }
 
         try await commandSender.setEnableGatewayIndependence(enableGatewayIndependence: true)
         guard case .needsRelaxedIndependenceCriteria = try await commandSender.getTentativeGateways()
