+++ nym-vpn-apple/NymVPNDaemon/Resources/Info.plist
@@ -36,7 +36,7 @@
 	<key>CFBundleVersion</key>
 	<string>$(CURRENT_PROJECT_VERSION)</string>
 	<key>IsCiBuild</key>
-	<string>$(IS_CI_BUILD)</string>
+	<string>true</string>
 	<key>LSUIElement</key>
 	<true/>
 	<key>NSHumanReadableCopyright</key>
