+++ nym-vpn-apple/Services/Sources/Services/ConnectionManager/ConnectionManager+macOS.swift
@@ -9,7 +9,9 @@ import WidgetKit
 
 extension ConnectionManager {
     @MainActor func connect() async throws {
-        try? await grpcManager.setGatewayIndependence(true)
+        // Reminders off means the user opted out of the warning: relax up front
+        // so the daemon never surfaces `needsRelaxedIndependenceCriteria` (no modal).
+        try? await grpcManager.setGatewayIndependence(appSettings.serverFamilyRemindersEnabled)
         try await startTunnel()
     }
